Reorg structure, move graph.js to data folder
