+++ src/demos/graph/data/sample-graph.js
